Multiple Fixes
- Move `bstrptr` to `test_strings` header as a common palceholder for all arch's
- Fix `iterations` to handle `uint64_t` datatype and larger values
- Replace the strings in `ptt2asm` after tokenization

+++ include/test_types.h
@@ -258,7 +258,7 @@ typedef struct {
 typedef struct {
     int help;
     int verbosity;
-    int iterations;
+    uint64_t iterations;
     double runtime;
     int csv;
     int json;
