+++ include/thread_group.h
@@ -35,12 +35,13 @@
 /*#include <string.h>*/
 /*#include <unistd.h>*/
 /*#include <map.h>*/
+#include <inttypes.h>
 #include <stdint.h>
 #include <pthread.h>
 
 #include "test_types.h"
 
-#define MIN_ITERATIONS 10
+#define MIN_ITERATIONS ((uint64_t)10)
 #define MIN_RUNTIME 1.0
 #define TIMEOUT_SECONDS 60
